RuntimeError
Expected object of scalar type Long but got scalar type Int for argument #2 'target' in call to _thnn_nll_1oss2d_forward, for this error we had to change the format of target into format torch.long

+++ models/architectures.py
@@ -351,7 +351,7 @@ class KPFCNN(nn.Module):
         """
 
         # Set all ignored labels to -1 and correct the other label to be in [0, C-1] range
-        target = - torch.ones_like(labels)
+        target = - torch.ones_like(labels, dtype=torch.long)
         for i, c in enumerate(self.valid_labels):
             target[labels == c] = i
